Drop the concept of a 'config file' - unecessary
In particular, the config file was included just to see how it felt, plus since we were already allowing the user to configure the 'TODAY_DIR' via an env var, why not also allow it through a config file. But the config file was unnecessary code, and the env var configuration we get for free from go-arg.